requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
google-genai==1.69.0
pdfplumber
pymongo
motor